+++ elasticsearch_dsl/search.py
@@ -97,7 +97,7 @@ class Request(object):
         Specify query params to be used when executing the search. All the
         keyword arguments will override the current values. See
         http://elasticsearch-py.readthedocs.org/en/master/api.html#elasticsearch.Elasticsearch.search
-        for all availible parameters.
+        for all available parameters.
 
         Example::
 
@@ -110,7 +110,7 @@ class Request(object):
 
     def index(self, *index):
         """
-        Set the index for the search. If called empty it will rmove all information.
+        Set the index for the search. If called empty it will remove all information.
 
         Example:
 
@@ -203,7 +203,7 @@ class Search(Request):
         :arg index: limit the search to index
         :arg doc_type: only query this type.
 
-        All the paramters supplied (or omitted) at creation type can be later
+        All the parameters supplied (or omitted) at creation type can be later
         overriden by methods (`using`, `index` and `doc_type` respectively).
         """
         super(Search, self).__init__(**kwargs)
@@ -503,7 +503,7 @@ class Search(Request):
 
     def highlight(self, *fields, **kwargs):
         """
-        Request highliting of some fields. All keyword arguments passed in will be
+        Request highlighting of some fields. All keyword arguments passed in will be
         used as parameters. Example::
 
             Search().highlight('title', 'body', fragment_size=50)
@@ -535,7 +535,7 @@ class Search(Request):
         All keyword arguments will be added to the suggestions body. For example::
 
             s = Search()
-            s = s.suggest('suggestion-1', 'Elasticserach', term={'field': 'body'})
+            s = s.suggest('suggestion-1', 'Elasticsearch', term={'field': 'body'})
         """
         s = self._clone()
         s._suggest[name] = {'text': text}